Introduce IBuyBaloot -> Ashkal, Hukom, Cut, Sun. Randomize RoomManager.IDs, move GameResult to its own file
