+++ src/main/java/Core/Ashkal.java
@@ -1,0 +1,7 @@
+package Core;
+
+
+public class Ashkal implements BalootBuy {
+
+}
+
