+++ src/main/java/Core/BalootBuy.java
@@ -1,0 +1,4 @@
+package Core;
+
+public interface BalootBuy {
+}
