+++ src/main/java/Core/Cut.java
@@ -1,0 +1,6 @@
+package Core;
+
+// Cutter, And Dealer, are 'something' general
+public class Cut {
+
+}
