+++ src/main/java/Core/Hukom.java
@@ -1,0 +1,5 @@
+package Core;
+
+public class Hukom implements BalootBuy {
+
+}
