+++ src/main/java/Core/Room.java
@@ -69,20 +69,3 @@ public class Room {
     }
 }
 
-class GameResult {
-//    everyRoom has a GameResult(Like a decider);
-    Room TheRoom;
-    int score; // If this is >= 152, room.finished==true.
-    public GameResult(Room theRoom) {
-        TheRoom = theRoom;
-    }
-
-    public boolean ShouldEndGame() {
-        if (score >= 152) {
-            TheRoom.Finished = true;
-            return true;
-        }
-        return false;
-    }
-
-}
