+++ src/main/java/Core/Sun.java
@@ -1,0 +1,4 @@
+package Core;
+
+
+public class Sun implements BalootBuy {}
